improvised Sidebar theme
