+++ client/src/components/Sidebar/Sidebar.css
@@ -1,0 +1,5 @@
+.sidebar-heading {
+  font-size: 0.75rem;
+  letter-spacing: 1px;
+  opacity: 0.8;
+}
